Update auth.js
this config constant is being used to store the JWT secret which is why in the const secret we have it pointing to config.SECRET because we are holding the JWT secret in a config.js file which is being ignored by .gitignore. I will speak with Austin about this change and update him accordingly.